Updated to support newest Gnome Terminal
The latest versions of Gnome Terminal store their settings and profiles
using the newer dconf system rather than gconf.  As such, the script
will now detect both the presence of dconf and the appropriate dconf
keys for Gnome Terminal.  When the dconf keys are not found, then it
will gracefully fall back to the old method using gconf.