Add "Xcode" directory
Create an "Xcode" directory and copy the current contents of the
"Xcode 4" directory to it.

Previously, the "Xcode 4" directory was the only Xcode-related
directory in the repo, which is misleading because the themes
contained in it also work fine in Xcode 6 (beta 5).

Other projects (e.g., Alcatraz) use URLs that link directly to the
existing themes in the "Xcode 4" directory, so instead of removing the
"Xcode 4" directory and breaking everything that might be linking to
those themes, I propose we leave it for now.

Moving forward, instead of supporting each version of Xcode by creating
an separate "Xcode <version>" directory -- which wasn't done for
Xcode 5 or Xcode 6 -- I propose there be a single "Xcode" directory
that supports the latest version of Xcode.