Fix registry key space encoding
The unescaped space in the registry key name for "Tomorrow Night" in PuTTY caused unexpected behavior. The theme would load, but attempting to save any additional settings the user needed to change (e.g. adding a SSH key file) would spawn a new settings entry with default colors. This commit changes the space character to `%20`, which shows up correctly as a space and allows the imported settings to be edited and saved as expected.